requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-xai>=0.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
flask>=3.0.0
twilio>=9.0.0
